# executor: Improve the performance of `aggFuncBitXor` by using sliding window