--- a/executor/aggfuncs/func_bitfuncs.go
+++ b/executor/aggfuncs/func_bitfuncs.go
@@ -85,6 +85,31 @@ func (e *bitXorUint64) UpdatePartialResult(sctx sessionctx.Context, rowsInGroup 
 	return nil
 }
 
+func (e *bitXorUint64) Slide(sctx sessionctx.Context, rows []chunk.Row, lastStart, lastEnd uint64, shiftStart, shiftEnd uint64, pr PartialResult) error {
+	p := (*partialResult4BitFunc)(pr)
+	for i := uint64(0); i < shiftStart; i++ {
+		inputValue, isNull, err := e.args[0].EvalInt(sctx, rows[lastStart+i])
+		if err != nil {
+			return err
+		}
+		if isNull {
+			continue
+		}
+		*p ^= uint64(inputValue)
+	}
+	for i := uint64(0); i < shiftEnd; i++ {
+		inputValue, isNull, err := e.args[0].EvalInt(sctx, rows[lastEnd+i])
+		if err != nil {
+			return err
+		}
+		if isNull {
+			continue
+		}
+		*p ^= uint64(inputValue)
+	}
+	return nil
+}
+
 func (*bitXorUint64) MergePartialResult(sctx sessionctx.Context, src, dst PartialResult) error {
 	p1, p2 := (*partialResult4BitFunc)(src), (*partialResult4BitFunc)(dst)
 	*p2 ^= uint64(*p1)