### What problem does this PR solve?

Problem Summary: Improve the performance of  the aggFuncBitXor


related #14887

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

- Unit test

Side effects

* benchmark
```
name                                                                                                                                    old time/op  new time/op  delta
rows
WindowFunctionsWithSlidingWindow/(func:bit_xor,_aggColType:int(11),_numFunc:1,_ndv:100,_rows:100000,_sorted:true,_concurrency:1)-16     21.8ms ± 1%   7.2ms ± 1%  -66.95%  (p=0.000 n=10+10)
ranges
WindowFunctionsWithSlidingWindow/(func:bit_xor,_aggColType:int(11),_numFunc:1,_ndv:100,_rows:100000,_sorted:true,_concurrency:1)#01-16   738ms ± 1%     8ms ± 2%  -98.95%  (p=0.000 n=9+10)
```

[new.txt](https://github.com/pingcap/tidb/files/4438883/new.txt)
[old.txt](https://github.com/pingcap/tidb/files/4438884/old.txt)